Taint detector refactored to a hierarchy of classes
